semi-automatic fixes

+++ src/main.rs
@@ -51,7 +51,7 @@ async fn handle_receiving(
         views::ask_lobby_id_and_join_lobby(&mut client, lobbies).await?;
     }
 
-    let mut selected_index = 0 as usize;
+    let mut selected_index = 0;
     loop {
         let game_mode = views::choose_game_mode(&mut client, &mut selected_index).await?;
         match game_mode {
